Added in JavaScript
Renamed the files to their proper names. Added in the date and my name into the comments. Began commenting the code. Added in the scripts to the html. Added in a script to get the content from another script and put it in the index.html. Added the HTML content from a ton of arrays into a variable. That variable is then added into the aside of the page.